refactor: add folder for tests files
